requirements.txt
asgiref==3.8.1
DateTime==5.5
Django==5.0.14
django-localflavor==1.6.2
django-locations-base==0.3.4
django-redis==6.0.0
djangorestframework==3.16.0
geographiclib==1.52
geopy==1.23.0
mssql-django==1.5
pillow==11.2.1
point==0.0.1
pyodbc==5.2.0
python-dotenv==1.1.1
pytz==2025.2
redis==6.2.0
setuptools==80.9.0
six==1.11.0
sqlparse==0.5.3
tzdata==2025.2
zope.interface==7.2
flake8